+++ rpc-augment/detectOther.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2022 @polkadot/rpc-augment authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { packageInfo as coreInfo } from 'https://deno.land/x/polkadot@0.2.0/rpc-core/packageInfo.ts';
-import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot@0.2.0/types/packageInfo.ts';
+import { packageInfo as coreInfo } from 'https://deno.land/x/polkadot/rpc-core/packageInfo.ts';
+import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot/types/packageInfo.ts';
 
 export default [coreInfo, typesInfo];