+++ types-known/detectOther.ts
@@ -1,6 +1,6 @@
 // Copyright 2017-2022 @polkadot/types-known authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot@0.2.0/types/packageInfo.ts';
+import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot/types/packageInfo.ts';
 
 export default [typesInfo];