+++ types-known/upgrades/manual/index.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2022 @polkadot/types-known authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as kusama } from './kusama.ts';
+export { default as polkadot } from './polkadot.ts';
+export { default as westend } from './westend.ts';